Improved FileButton behavior, opening the same file twice in a row will now work. Also simplified its implementation.

+++ deploy.ps1
@@ -3,9 +3,11 @@
 
 Write-Output "Running tests."
 yarn test
+if ($LastExitCode -ne 0) { throw "Tests failed." }
 
 Write-Output "Generating production build."
 yarn build
+if ($LastExitCode -ne 0) { throw "Build failed." }
 
 Write-Output "Deleting ./deployment contents."
 Remove-Item -Recurse ./deployment/*
@@ -22,9 +24,13 @@ Write-Output $version > version.txt
 Write-Output "Committing and pushing to gh-pages."
 Set-Location ./deployment
 git pull
+if ($LastExitCode -ne 0) { throw "Git pull failed." }
 git add .
+if ($LastExitCode -ne 0) { throw "Git add failed." }
 git commit -m "Release $version."
+if ($LastExitCode -ne 0) { throw "Git commit failed." }
 git push
+if ($LastExitCode -ne 0) { throw "Git push failed." }
 
 Set-Location ..
